Extract default status bar module
